refactor(inspection-service): extract inspection-results sub-service
Move updateResults / getPropertyFacts / updatePropertyFacts /
updateTemplateSnapshot / switchRatingSystem into InspectionResultsService.
Self-contained (getPropertyFacts is called internally by updatePropertyFacts).
The facade delegates; moved code is byte-identical (dynamic schema imports
re-pathed for the new module depth).
